Cleanup: Moved core classes to the JAPI namespace
